мелкие исправления
