--- src/util/tools.js
@@ -1,2 +1,0 @@
-export const getRandomBool = (chance = 0.5) =>
-  Math.random() > chance;
